refactor: modal refactoring
